Work to re-organize project better and allow easier support of future ideas.
