+++ BragiFont.Core.OLD/Internal/Character.cs
