+++ BragiFont.Core.OLD/Internal/GlyphCache.cs
