+++ BragiFont.Core.OLD/Internal/TextCharacter.cs
