+++ BragiFont.Core.OLD/Internal/TextImplementation.cs
