+++ BragiFont.Core.OLD/Text.cs
